move ProgressBars from run.py to utilities.py
